updating choose symbol code
