+++ src/scss/header.scss
@@ -4,7 +4,7 @@
     color: rgb(49, 9, 26)
 }
 
-.startWrapper {
+.resetWrapper {
     display: flex;
     justify-content: center;
     padding-top: 2rem;
